Move EditorWindow to the ShaderGraph.GraphUI namespace.
