+++ com.unity.shadergraph/Editor/GraphUI/ShaderGraphMainToolbar.cs
@@ -1,6 +1,4 @@
-using System;
 using UnityEditor.GraphToolsFoundation.Overdrive;
-using UnityEngine;
 
 namespace UnityEditor.ShaderGraph.GraphUI
 {
